requirements.txt
mwclient
opencc
python-dotenv
